backend/services/recurring-task-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pydantic==2.5.3
pydantic-settings==2.1.0
dapr==1.12.0
cloudevents==1.10.1
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
structlog==24.1.0
prometheus-client==0.19.0
